CompiledPermutations.jl

CompiledPermutations is a micro-package which provides a function compile_permutation! that compiles a given permutation so that it can be efficiently applied multiple times using permute!. This is useful if you would like to permute several AbstractVectors in the same manner. For example, when sorting a DataFrame, StructArray, or similar data structure.

CompiledPermutations.compile_permutation! — Function

compile_permutation!(p::AbstractVector{<:Integer}) -> cp

Convert a permutation p into a compiled permutation cp which can be efficiently applied with permute!(v, cp).

permute!(v, compiled_permutation!(p))) produces the same result as permute!(v, p) and is often less efficient. This function is useful when the same permutation is applied to many vectors.

Note

compile_permutation! may mangle p and/or return an object which shares memory with p. Use compile_permutation!(copy(p)) if you would like to avoid this.

Example

julia> cp = compile_permutation!([2, 3, 1]);

julia> v = [:a, :b, :c];

julia> permute!(v, cp)
3-element Vector{Symbol}:
 :b
 :c
 :a

Here is an example where this function provides performance benefits:

julia> perm = randperm(1000); data = [rand(1000) for _ in 1:30];

julia> @btime for x in $data; permute!(x, $perm); end
  25.917 μs (30 allocations: 238.12 KiB)

julia> @btime (cp = compile_permutation!(copy($perm)); for x in $data; permute!(x, cp); end)
  16.708 μs (2 allocations: 15.88 KiB)

CompiledPermutations.compile_inverse_permutation! — Function

compile_inverse_permutation!(p::AbstractVector{<:Integer}) -> cp

Equivalent to compile_permutation!(invperm(p)), though possibly more efficient.